requirements.txt
absl-py==2.0.0
astunparse==1.6.3
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
deepface==0.0.80
Deprecated==1.2.14
dill==0.3.7
# dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
faiss-cpu==1.7.4
filelock==3.13.1
fire==0.5.0
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
gast==0.4.0
gdown==4.7.1
gevent==23.9.1
google-auth==2.26.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.0
gunicorn==21.2.0
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.13.1
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
mtcnn==0.1.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
pathlib==1.0.1
pillow==10.2.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
six==1.16.0
soupsieve==2.5
# tensorboard==2.13.0
# tensorboard-data-server==0.7.2
# tensorflow==2.13.1
# tensorflow-cpu==2.13.1
# tensorflow-estimator==2.13.0
# tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tqdm==4.66.1
typing_extensions==4.5.0
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
